refactor(common): common/schedule 패키지를 entity/enums/repository/service로 재정리
패키지 루트에 엔티티/enum/리포지토리/서비스 계층 클래스가 뒤섞여 있던 것을,
프로젝트 전반의 도메인 패키지 컨벤션(entity/repository/enums/service/...)에
맞춰 하위 패키지로 분리한다. 동작 변화는 없다(패키지 이동 + import 정리 +
ScheduledTask 필드별 Javadoc 보강).
