+++ src/main/java/com/remake/gone/common/schedule/service/ScheduledTaskAdminService.java
@@ -2,12 +2,12 @@ package com.remake.gone.common.schedule.service;
 
 import com.remake.gone.common.exception.CustomException;
 import com.remake.gone.common.response.PageResponse;
-import com.remake.gone.common.schedule.ScheduledTask;
-import com.remake.gone.common.schedule.ScheduledTaskRepository;
-import com.remake.gone.common.schedule.ScheduledTaskStatus;
 import com.remake.gone.common.schedule.dto.ScheduledTaskResponse;
 import com.remake.gone.common.schedule.dto.ScheduledTaskStatsResponse;
+import com.remake.gone.common.schedule.entity.ScheduledTask;
+import com.remake.gone.common.schedule.enums.ScheduledTaskStatus;
 import com.remake.gone.common.schedule.exception.ScheduleErrorCode;
+import com.remake.gone.common.schedule.repository.ScheduledTaskRepository;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import lombok.RequiredArgsConstructor;
