+++ src/main/java/com/remake/gone/common/schedule/service/ScheduledTaskService.java
@@ -1,5 +1,8 @@
-package com.remake.gone.common.schedule;
+package com.remake.gone.common.schedule.service;
 
+import com.remake.gone.common.schedule.entity.ScheduledTask;
+import com.remake.gone.common.schedule.enums.ScheduledTaskStatus;
+import com.remake.gone.common.schedule.repository.ScheduledTaskRepository;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.Optional;
